MS: cut down some nested styled components
